server/requirements.txt
django
djangorestframework
requests
numpy
cython
cmake
Pillow
dlib
face_recognition
imutils
opencv-python